// packages/shared/src/enums/DamageType.ts
export enum DamageType {
  BARREL_EXPLOSION = 'barrel_explosion',
  PROJECTILE_HIT = 'projectile_hit',
  ZONE_DAMAGE = 'zone_damage',
  TRAP_DAMAGE = 'trap_damage',
  MELEE_HIT = 'melee_hit',
  SUDDEN_DEATH = 'sudden_death',
  SIEGE_CRUSH = 'siege_crush',
  RANGED_HIT = 'ranged_hit',
  THROWN_HIT = 'thrown_hit',
}

export default DamageType;